requirements.txt
simplejson>=3.16.0
requests>=2.18.4
pytest==5.3.5
allure-pytest==2.8.18
pytest-ordering==0.6
configparser>=3.5.0
pytest-rerunfailures==5.0
protobuf>=3.13.2
xmltodict>=0.12.0
urllib3>=1.24.2
pymongo>=3.10.1
pycryptodome>=3.9.8
